+++ scripts/visualize.py
@@ -10,7 +10,7 @@ import argparse
 import pandas as pd
 import os
 
-from scripts.data_loader import load_network_data
+from scripts.data_utils import load_network_data
 from scripts.viz_utils import *
 
 #visualize the map
